[FIX] hr_timesheet_attendance: convert field type
Bad migration of new API, was a float field.
Probably introduced at 510d260

Closes #16290

+++ addons/hr_timesheet_attendance/models/hr_timesheet_sheet.py
@@ -10,7 +10,7 @@ class HrTimesheetSheet(models.Model):
     _inherit = "hr_timesheet_sheet.sheet"
 
     attendances_ids = fields.One2many('hr.attendance', 'sheet_id', 'Attendances')
-    total_attendance = fields.Integer(string='Total Attendance', compute='_compute_total')
+    total_attendance = fields.Float(string='Total Attendance', compute='_compute_total')
     total_timesheet = fields.Float(string='Total Timesheet', compute="_compute_total")
     total_difference = fields.Float(string='Difference', compute="_compute_total")
     period_ids = fields.One2many('hr_timesheet_sheet.sheet.day', 'sheet_id', string='Period', readonly=True)
